+++ correlate/src/data/column_config.rs
@@ -7,5 +7,7 @@ pub struct ColumnConfig {
     pub column_type: ColumnType,
     pub is_sortable: bool,
     #[serde(default)]
+    pub is_key: bool,
+    #[serde(default)]
     pub width: Option<f32>,
 }